Flesh out TestGrid
by adding fetchData(), and a nav controller, and making it
the root VC of acorn.  Rename to MainGrid.
